Merge vagrant-kubeadm-testing for contributor experience
We discussed on last week's kubeadm call, that we would like move this
repo into the kubeadm contributor repo, so we improve our shared tooling
and contributor experience.

This vagrant environment has been shown to aide in developer ramp up for
new contributors.
The documentation fixes provide a happier path to discovering how to
build the software.

approved by @timothysc

Fixes: https://github.com/stealthybox/vagrant-kubeadm-testing/issues/1