kinder: avoid polution of the output of "kubeadm config images list"
In etcd.go when we try to get the propper version of etcd
to use from "kubeadm config images list", if the version
of kubeadm is much older than "stable-1" (which is the default label
in kubeadm), the ouput shows a warning to stderr e.g.:

  I0626 02:54:47.931193    3673 version.go:240] remote version
  is much newer: v1.15.0; falling back to: stable-1.14

For some reason this output becomes part of the "etcdVersionBuf"
which is supposed to only contain stdout (cmd.SetStdout(&etcdVersionBuf)).

Omit this ouput by piping stderr to /dev/null and only
catch the etcd image string - e.g.:

  k8s.gcr.io/etcd:3.3.10

Once kubeadm supports machine readable output we can parse
this output properly without workarounds.