kinder: use pkg/config for managing the kind configuration
In recent versions of kind the internal types have moved under
"/internal" sub-packages which do not allow import.

This means that kinder cannot import them.

The purpose of the new kinder pkg/config package is to wrap the public
kind types and allow validation and defaulting on the public types
instead of forking the internal kind types.

The validation and defaulting is performed by forked utilities under
third_party/kind/config.