kinder: don't use "docker cp /dev/stdin"
Similar flakes were seen in kind when the command "export logs"
is used. The command also uses "/dev/stdin" for copying between
host and container. As much as being a speculation the "/dev/stdin"
flakes could be a Docker bug.

Create an utility function Node#WriteFile that instead of
using /dev/stdin uses a temporary file on the host and copies
the contents of the file to the node.

Use this function in a number of locations instead of using
"docker cp /dev/stdin ...".