kinder: pre-pull and import non-kube images during node-image alter
The idea of this change is to reduce the amount of network traffic
kinder e2e test workflows have to do with the attempt to fix flakes.

For 3CP setup etcd has to be pulled 3x~200MB (6x~200MB for upgrade).
After this change it has to be pulled 1x~200MB (2x~200MB for upgrade).

- pre-pull non-kube images for docker and containerd based on the output
of the kubeadm binary for init/join and binary for upgrade.
- pre-load these images for docker during cluster create and upgrade
- for containerd the images are not pre-loaded due to issues related
to the containerd "snapshoter" functionality. add TODOs to resolve
that in the future. one potential fix is to import them on runtime
the same way it is done for docker.
NOTE: currently we don't have e2e tests for containerd.

CNI images (Calico) are not pre-pulled.