kinder: avoid orphaning child processes on task timeouts
When a task has timed out, ideally the whole process tree should be
terminated with a SIGKILL.

The e2e suite has some caveats with AfterSuite() functions and
these can only contribute noise to the logs. A SIGKILL to
the parent process (Golang command) and the child processes such
as Ginkgo, ensure that everything is terminated right away.

Otherwise, Ginkgo can continue running while the cluster is
being destroyed and fill the logs with errors.

Originally this code was disabled for debugging, but executing it
is mandatory. Add more debug messages to help find problems around
obtaining a pgid or killing a process tree in case failures in CI
are seen again.