kinder: make the task runner perform cleanup only on existing processes
If the process PID no longer exists we can omit killing the process
tree in cleanup(). Passing 0 as the signal to syscall.Kill() is used
to determine if the process is still running.

https://man7.org/linux/man-pages/man2/kill.2.html

"If sig is 0, then no signal is sent, but existence and permission
checks are still performed; this can be used to check for the
existence of a process ID or process group ID that the caller is
permitted to signal."