use k8s-release-dev instead of kubernetes-release-dev
this was a straight search-replace rather than attempting to converge
toward dl.k8s.io because way too much of this is stale. I tried editing
to update stale info at first, but someone who knows the state of things
needs to do this.

e.g. many of the uris in here are already dead, let alone things like
/bazel and /ci-cross not existing in gs://k8s-release-dev