kinder: fix failing patches e2e tests
- Looks like /tmp is not a good location to copy files to.
Either they silently fail to copy to /tmp or they throw an
error that a /tmp/sub-dir like "kinder-patches" is not a directory
(but it is). Searching for the related errors online:
  - Error response from daemon: extraction point is not a directory
  - Error: No such container:path: kinder-patches-control-plane-1:/tmp/kubeadm-patches
does not provide much context on why this is happening.
Use "/kinder/patches" instead of "/tmp...".

- Include missing error return in copyPatchesToNode()

- Remove host OS check about the patch directory. We don't need this check
because an earlier step should copy the patches from host to container.
If the patch directory in the container is different from the patch
directory on the host it would fail.

- Given patches.directory is always part of the kubeadm config,
always create the patch directory with copyPatchesToNode(). Skip
copying patches if the source directory is not defined.