github/workflows/support-command: don't error on comments with no command
If the commenting GitHub users pass a comment without the "/support"
command the GitHub action will throw an error.

This seems like a missing feature in the xt0rted/slash-command-action
dependency. Include a workaround that skips the comment/close/label
steps if the command is not present.

Also pin to ubuntu-20.04 instead of latest.