kinder: switch all workflows to containerd
Switch all e2e test workflows to use containerd as the CR for the time
being. Dockershim is being removed from k/k and until cri-dockerd
gains maintainer traction we should avoid setting up our selfs
for the sake of testing Docker as the CR.

Technically the dockershim removal only affects "latest" workflows
but this change will apply it to all k/k branches. Once and if
cri-dockerd is testable we can add a single workflow for Docker
as the CR.