status: use a podman-compatible format string
Both `docker` and `podman` expose the `.Labels` map, but only `docker`
exposes a method on the `ps` output to index into the map. This method
is superfluous, since Go templates let you index into the map natively.
Using the native indexing allows the `--format` string to work for
`docker` and `podman` both.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>