kinder: optimize the preparation of upgrade artifacts
During the "altering" of the node image used by kinder,
there are two two download stages, one to download the init
artifacts and one to download upgrade artifacts.

For some e2e workflows we have matching init and upgrade versions.
This is done just to test the upgrade command and not perform
an actual upgrade. It causes the same artifacts to be downloaded
two times, which is slow.

During the preparation of bits installers, check if the init/upgrade
versions match. If so, tell the upgrade bits installer to use
the init directory as install source. This results in just
copying the bits from the host, instead of downloading them
again.