kinder: call "upgrade plan" before "upgrade apply"
During the kinder "kubeadm-upgrade" action, "upgrade apply"
is executed, but "upgrade plan" is not. To ensure better coverage
ensure "upgrade plan" is called as well. We have this command
as part of the kubeadm upgrade docs, so it makes sense to have
it here.

Ensure that "plan" is also called during the
"upgrade-latest-no-addon-config-maps" workflow.