kinder: remove parallel code during node creation
Running docker commands in parallel has proven to confuse
the docker CR on the host and cause random failures.
Also it is less clear what container command failed exactly.

Make the whole node creation linear. This will inevitably make
the kinder cluster creation slower, but will reduce
the change for unexplainable flakes and allow us to debug
and potentially retry certain commands that are flaky.

In the past one such command was a simple:
    "docker exec ... mkdir ..."