kinder: support patching OCI image tar archives
When kinder is building a node image it fetches image
TAR files for the k8s components from the CI bucket.
These image TARs contain spec tagged as '-amd64", so
they are patched to not have this suffix before
being imported in containerd. This makes them compatible
with kubeadm commands.

Recently the image format changed from Docker to OCI.
To support OCI, parse an additional "index.json" file
which is part of the OCI image spec.

Docker image TARs continue to be supported.